// demohouse/storybook-agent/frontend/src/common/components/StoryBook/hooks/useResponsiveState/interface.ts
/*
 * Copyright (c) 2025 Bytedance Ltd. and/or its affiliates
 * Licensed under the 【火山方舟】原型应用软件自用许可协议
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at 
 *     https://www.volcengine.com/docs/82379/1433703
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export enum strategyEnum {
  matchFirst = 'matchFirst',
  matchLast = 'matchLast',
}

export interface Options<T> {
  defaultValue?: T
  strategy?: strategyEnum
}
export interface MediaQueryList {
  matchMedia: string
}

export interface MatchMediaList<T> {
  value: T
  matchMedia: string
}

export interface MatchMediaListSimple<T> {
  value: T
  min: number | string
  max: number | string
}